requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.2
django-ckeditor==6.7.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-multiupload==0.6.1
idna==3.6
Pillow==10.1.0
requests==2.31.0
sqlparse==0.4.4
sslcommerz-lib==1.0
tzdata==2023.3
urllib3==2.2.1
whitenoise==6.6.0
